refactor

+++ agent/tests/fixtures.py
@@ -17,9 +17,10 @@ def cli_runner():
     for filename in os.listdir(pipeline_cli.SDC_DATA_PATH):
         if filename.startswith('error-test_'):
             os.remove(os.path.join(pipeline_cli.SDC_DATA_PATH, filename))
-    for filename in os.listdir(pipeline_cli.SDC_RESULTS_PATH):
-        if filename.startswith('sdc-test_'):
-            os.remove(os.path.join(pipeline_cli.SDC_RESULTS_PATH, filename))
+    if os.path.isdir(pipeline_cli.SDC_RESULTS_PATH):
+        for filename in os.listdir(pipeline_cli.SDC_RESULTS_PATH):
+            if filename.startswith('sdc-test_'):
+                os.remove(os.path.join(pipeline_cli.SDC_RESULTS_PATH, filename))
 
     if os.path.isfile(pipeline_cli.TOKEN_FILE):
         os.remove(pipeline_cli.TOKEN_FILE)
